feat: [#4] 웍스 설정 카드 분리
